refactor : 패키지, 파일명 변경 (#616)
